Generalize multiprocessing shared array for labels images
Refactor generating and converting shared memory arrays for multiprocessing of labels images into a separate class. Also, generalize the initialization function.
